NBAX-250 #resolve #comment Changed id and sourceSystemId to record uri
Because the taxon ids provided by CoL (may) change each year, this
update now uses the id part from the record uri as stable id. This id is
being used as basis for the document id (ES id) as well as the
sourceSystemID. NOTE that this change only effects the json export
(CoLTaxonFullImporter); the importer that loads the documents directly
into the document store (CoLTaxonImporter) has not been changed.