NBAX-253 #resolve #comment Fix for always true | always false condition
The ConditionValidator and ConditionCollector didn't handle the
FalseCondition and the TrueCondition correctly. This has been fixed. At
the same time the FalseCondition has been modified, so that it no longer
returns a boolQuery (i.e. mustNot(matchAllQuery), but a
MatchNoneQueryBuilder instead.

Unit tests have been updated to demonstrate that translations of these
conditions match the expectations.